todo-service/requirements.txt
alembic
bcrypt
fastapi[standard]
sqlmodel
pydantic-settings
pyjwt
pytest
pytest-dotenv
psycopg2
httpx